+++ ChangeLog.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2022-05-08
+### Added
+- Discount Created Event
+- Discount Created Listener
+
+
+
 ## [0.0.2] - 2022-05-08
 ### Added
 - BootUpSeeder to build the main objects